feat: moved floor test-cases to the test package and added a new testcase
